consolidate routing hooks
