+++ app/src/organisms/ErrorRecoveryFlows/hooks/useRecoveryRouting.ts
@@ -2,7 +2,7 @@ import * as React from 'react'
 
 import { RECOVERY_MAP } from '../constants'
 
-import type { IRecoveryMap } from '../types'
+import type { IRecoveryMap, RecoveryRoute } from '../types'
 import type { ERUtilsResults } from './useERUtils'
 
 /**
@@ -15,6 +15,7 @@ import type { ERUtilsResults } from './useERUtils'
 
 export function useRecoveryRouting(): {
   recoveryMap: IRecoveryMap
+  previousRoute: RecoveryRoute | null
   setRM: (map: IRecoveryMap) => void
   trackExternalMap: ERUtilsResults['trackExternalMap']
 } {
@@ -26,9 +27,27 @@ export function useRecoveryRouting(): {
   // If we do multi-app routing, concat the sub-step to the error recovery routing.
   const [, setSubMap] = React.useState<Record<string, any> | null>(null)
 
+  const previousRoute = usePreviousRecoveryRoute(recoveryMap.route)
+
   return {
     recoveryMap,
+    previousRoute,
     setRM: setRecoveryMap,
     trackExternalMap: setSubMap,
   }
+}
+
+// The previous recovery route visited before the current route, if any.
+export function usePreviousRecoveryRoute(
+  route: RecoveryRoute
+): RecoveryRoute | null {
+  const [prevRoute, setPrevRoute] = React.useState<RecoveryRoute | null>(null)
+  const [currentRoute, setCurrentRoute] = React.useState<RecoveryRoute>(route)
+
+  if (route !== currentRoute) {
+    setPrevRoute(currentRoute)
+    setCurrentRoute(route)
+  }
+
+  return prevRoute
 }
